Add files via upload
This program allows the user to select food items from a list and calculates the total nutritional value of the items chosen. The user is presented with a list of food options, such as:

1: Apple
2: French Fries
3: Burger
4: Instant Ramen
5: Protein Bar
...
11: Finished

The program will continue looping, allowing the user to select multiple food items, until the user selects "Finished." Once "Finished" is chosen, the program will calculate and display the total nutritional value of all selected items and provide some nutritional statistics.